+++ lib/src/api/devices.dart
@@ -227,7 +227,7 @@ Future<List<NativeMediaStreamTrack>> _getUserMediaFFI(
       constraints.audio.mandatory != null || constraints.audio.optional != null
           ? ffi.AudioConstraints(
             deviceId: constraints.audio.mandatory?.deviceId,
-            processing: ffi.AudioProcessingConfig(
+            processing: ffi.AudioProcessingConstraints(
               autoGainControl:
                   constraints.audio.mandatory?.autoGainControl ??
                   constraints.audio.optional?.autoGainControl,
@@ -317,7 +317,7 @@ Future<List<NativeMediaStreamTrack>> _getDisplayMediaFFI(
       constraints.audio.mandatory != null || constraints.audio.optional != null
           ? ffi.AudioConstraints(
             deviceId: constraints.audio.mandatory?.deviceId,
-            processing: ffi.AudioProcessingConfig(),
+            processing: ffi.AudioProcessingConstraints(),
           )
           : null;